Create base version of vmaas-webapp-next
It's a re-implementation of the webapp container in go, attempting
to reduce resource requirements of this part of vmaas.

It uses new dump format based on small sqlite database, and utilizes libraries
used in patchman engine to provide same api as the original python version.
The current implementation only contains code for the `/updates` call,
and proxies all other requests to original container.
There are some idiosyncratic patterns in this new version to preserve absolute
compatibility, namely using `*[]string` to store `available_updates`,
since it can be missing, or an empty array depending uf the input package
is a valid NEVRA & whether the name is found in the database.

The current implementation exhibits reduced memory requirements, and these can
be lowered even more by specifying GOGC env variable to ensure more
aggresive garbage collection.